services/authentication/requirements.txt
fastapi[standard]==0.115.11
pydantic==2.10.6
pydantic-settings==2.8.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
sqlmodel==0.0.22
mysqlclient==2.2.7